added basic player death
