+++ Assets/Scenes/Persistent.unity
@@ -938,6 +938,8 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   fadeTime: 0.25
+  defaultLevel: {fileID: 11400000, guid: bcc460598eae4314a9bba6eeca8b8739, type: 2}
+  playerPrefab: {fileID: 1831558400415614749, guid: 9d42730b7845e9b4ca5f64f4d2cb3a4e, type: 3}
 --- !u!1 &1746217393
 GameObject:
   m_ObjectHideFlags: 0
